+++ src/main/java/com/verdantartifice/primalmagic/common/spells/payloads/FrostDamageSpellPayload.java
@@ -86,7 +86,7 @@ public class FrostDamageSpellPayload extends AbstractDamageSpellPayload {
     }
 
     protected int getDurationSeconds(SpellPackage spell, ItemStack spellSource) {
-        return this.getModdedPropertyValue("duration", spell, spellSource);
+        return 2 * this.getModdedPropertyValue("duration", spell, spellSource);
     }
 
     @Override