+++ src/main/java/com/verdantartifice/primalmagic/common/spells/payloads/SolarDamageSpellPayload.java
@@ -89,7 +89,7 @@ public class SolarDamageSpellPayload extends AbstractDamageSpellPayload {
     }
 
     protected int getDurationSeconds(SpellPackage spell, ItemStack spellSource) {
-        return this.getModdedPropertyValue("duration", spell, spellSource);
+        return 2 * this.getModdedPropertyValue("duration", spell, spellSource);
     }
 
     @Override